+++ CHANGELOG.md
@@ -17,13 +17,16 @@
 - `image`
   - [75336e34](https://github.com/wix-incubator/rich-content/commit/75336e34) open imageStudio on crop tab
 - `gallery`
- - [20e2af4](https://github.com/wix-incubator/rich-content/commit/20e2af4) height issues - disabled above the fold 
+ - [20e2af4](https://github.com/wix-incubator/rich-content/commit/20e2af4) height issues - disabled above the fold
+ - [#401](https://github.com/wix-incubator/rich-content/pull/401) resize media url returns absolute url when not resizing
 - `example`
   - [9add0fbc](https://github.com/wix-incubator/rich-content/commit/9add0fbc) gallery validation schema used for content validation
   - [#395](https://github.com/wix-incubator/rich-content/pull/395) removed div wrapping editor that was added in [#359](https://github.com/wix-incubator/rich-content/pull/359)
 ### :house: Internal
 - `e2e`
   - [#393](https://github.com/wix-incubator/rich-content/pull/393) added atomic block alignment testing
+- `gallery`
+  - [#401](https://github.com/wix-incubator/rich-content/pull/401) `pro-gallery` updated to `v1.3.17`
 <hr/>
 
 ## 5.1.1 (Sep 22, 2019)